\encoding{UTF-8}
\name{deque-package}
\alias{deque-package}
\alias{deque}
\docType{package}
\title{
\packageTitle{deque}
}
\description{
\packageDescription{deque}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{deque}
\packageIndices{deque}

}
\author{
\packageAuthor{deque}

Maintainer: \packageMaintainer{deque}
}
\references{
	Donald Knuth. The Art of Computer Programming, Volume 1: Fundamental Algorithms, Third Edition. Addison-Wesley, 1997. ISBN 0-201-89683-4. Section 2.2.1: Stacks, Deques, and Dedeques, pp. 238–243
}
\keyword{ package }
\examples{
	x <- deque()
	str(x)
}
